Rename and cleanup
