--- ansible_gatekeeper/__version__.py
@@ -1,1 +1,0 @@
-__version__ = "v0.0.1"
