--- ansible_gatekeeper/policybook/image/flow.drawio
@@ -1,64 +1,0 @@
-<mxfile host="65bd71144e">
-    <diagram id="RML1NZcLBZM324xWg7Dt" name="Page-1">
-        <mxGraphModel dx="784" dy="727" grid="1" gridSize="10" guides="1" tooltips="1" connect="1" arrows="1" fold="1" page="1" pageScale="1" pageWidth="827" pageHeight="1169" math="0" shadow="0">
-            <root>
-                <mxCell id="0"/>
-                <mxCell id="1" parent="0"/>
-                <mxCell id="18" value="" style="rounded=0;whiteSpace=wrap;html=1;strokeColor=#d6b656;strokeWidth=2;fillColor=#fff2cc;opacity=70;" vertex="1" parent="1">
-                    <mxGeometry x="190" y="110" width="450" height="140" as="geometry"/>
-                </mxCell>
-                <mxCell id="7" style="edgeStyle=none;html=1;entryX=0;entryY=0.5;entryDx=0;entryDy=0;entryPerimeter=0;" parent="1" source="2" target="5" edge="1">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="2" value="" style="sketch=0;pointerEvents=1;shadow=0;dashed=0;html=1;labelPosition=center;verticalLabelPosition=bottom;verticalAlign=top;align=center;shape=mxgraph.mscae.intune.policy;fillColor=#dae8fc;strokeColor=#6c8ebf;" parent="1" vertex="1">
-                    <mxGeometry x="100" y="157" width="50" height="50" as="geometry"/>
-                </mxCell>
-                <mxCell id="8" style="edgeStyle=none;html=1;" parent="1" source="5" target="6" edge="1">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="5" value="" style="verticalLabelPosition=bottom;html=1;verticalAlign=top;align=center;shape=mxgraph.azure.code_file;pointerEvents=1;fillColor=#007FFF;strokeColor=#007FFF;" parent="1" vertex="1">
-                    <mxGeometry x="263.75" y="157" width="47.5" height="50" as="geometry"/>
-                </mxCell>
-                <mxCell id="15" style="edgeStyle=none;html=1;exitX=1;exitY=0.5;exitDx=0;exitDy=0;exitPerimeter=0;" edge="1" parent="1" source="6" target="14">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="6" value="" style="sketch=0;aspect=fixed;pointerEvents=1;shadow=0;dashed=0;html=1;strokeColor=none;labelPosition=center;verticalLabelPosition=bottom;verticalAlign=top;align=center;fillColor=#00188D;shape=mxgraph.mscae.enterprise.code_file" parent="1" vertex="1">
-                    <mxGeometry x="400" y="157" width="49" height="50" as="geometry"/>
-                </mxCell>
-                <mxCell id="9" value="Policybook" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;fontStyle=1" vertex="1" parent="1">
-                    <mxGeometry x="80" y="212" width="90" height="30" as="geometry"/>
-                </mxCell>
-                <mxCell id="10" value="AST" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;fontStyle=1" vertex="1" parent="1">
-                    <mxGeometry x="263.75" y="207" width="50" height="30" as="geometry"/>
-                </mxCell>
-                <mxCell id="11" value="Rego" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;fontStyle=1" vertex="1" parent="1">
-                    <mxGeometry x="400" y="207" width="50" height="30" as="geometry"/>
-                </mxCell>
-                <mxCell id="12" value="to_ast.py&lt;br&gt;" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;" vertex="1" parent="1">
-                    <mxGeometry x="188.75" y="177" width="70" height="30" as="geometry"/>
-                </mxCell>
-                <mxCell id="13" value="to_rego.py" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;" vertex="1" parent="1">
-                    <mxGeometry x="310" y="177" width="80" height="30" as="geometry"/>
-                </mxCell>
-                <mxCell id="21" value="" style="edgeStyle=none;html=1;" edge="1" parent="1" source="14" target="20">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="14" value="Ansible Gatekeeper&lt;br&gt;Engine" style="rounded=1;whiteSpace=wrap;html=1;strokeColor=#b85450;fillColor=#f8cecc;" vertex="1" parent="1">
-                    <mxGeometry x="490" y="152" width="120" height="60" as="geometry"/>
-                </mxCell>
-                <mxCell id="17" style="edgeStyle=none;html=1;entryX=0.5;entryY=0;entryDx=0;entryDy=0;" edge="1" parent="1" source="16" target="14">
-                    <mxGeometry relative="1" as="geometry"/>
-                </mxCell>
-                <mxCell id="16" value="Project" style="whiteSpace=wrap;html=1;shape=mxgraph.basic.document;strokeColor=#666666;fillColor=#f5f5f5;strokeWidth=2;fontColor=#333333;" vertex="1" parent="1">
-                    <mxGeometry x="520" y="20" width="60" height="65" as="geometry"/>
-                </mxCell>
-                <mxCell id="20" value="evaluation&lt;br&gt;result" style="whiteSpace=wrap;html=1;shape=mxgraph.basic.document;strokeColor=#666666;fillColor=#f5f5f5;strokeWidth=1;fontColor=#333333;" vertex="1" parent="1">
-                    <mxGeometry x="690" y="142" width="70" height="80" as="geometry"/>
-                </mxCell>
-                <mxCell id="22" value="Ansible Gatekeeper" style="text;html=1;align=center;verticalAlign=middle;resizable=0;points=[];autosize=1;strokeColor=none;fillColor=none;" vertex="1" parent="1">
-                    <mxGeometry x="190" y="110" width="130" height="30" as="geometry"/>
-                </mxCell>
-            </root>
-        </mxGraphModel>
-    </diagram>
-</mxfile>
