+++ ansible_policy/__version__.py
@@ -1,0 +1,1 @@
+__version__ = "v0.0.1"
