--- examples/ansible-gatekeeper.cfg
@@ -1,9 +1,0 @@
-[policy]
-default disabled
-policies.community.*      tag=security    enabled
-policies.org.compliance   tag=compliance  enabled
-
-[source]
-# policies.community.mongodb = policies.community_mongodb:0.0.1     # collection policy (ansible-galaxy)
-# policies.community.mongodb = examples/policies-community_mongodb-0.0.1.tar.gz   # collection policy (local tarball)
-policies.org.compliance    = ../ansible-demo-policies    # org-wide compliance policy
