+++ examples/ansible-policy.cfg
@@ -1,0 +1,7 @@
+[policy]
+default disabled
+policies.community.*      tag=security    enabled
+policies.org.compliance   tag=compliance  enabled
+
+[source]
+policies.org.compliance    = ../ansible-demo-policies    # org-wide compliance policy
