--- examples/old/collection_policies/policies.community_mongodb/README.md
@@ -1,3 +1,0 @@
-# Ansible Collection - policies.community_mongodb
-
-Documentation for the collection.
