--- examples/old/collection_policies/policies.community_mongodb/policies/check_database_name.yml
@@ -1,36 +1,0 @@
-- hosts: localhost
-  vars:
-    allowed_databases: [] # <== should be input at runtime
-    package_name: "mongodb_user_db_policy"
-    filepath: "/tmp/{{ package_name }}.rego"
-  tasks:
-    - gatekeeper.rego.vars:
-        vars:
-          _allowed_databases: "{{ allowed_databases }}"
-      register: declare_vars
-
-    - gatekeeper.rego.rule:
-        name: database
-        exprs:
-          - db_name := resolve_var(input._agk.task.module_options.database, input._agk.task)
-        return: db_name
-      register: database
-
-    - gatekeeper.rego.check:
-        name: using_forbidden_database
-        exprs:
-          - database != ""
-          - not database in _allowed_databases
-      register: using_forbidden_database
-
-    - gatekeeper.rego.transform:
-        filepath: "{{ filepath }}"
-        package_name: "{{ package_name }}"
-        tags:
-          - security
-        match:
-          - module: "community.mongodb.mongodb_user"
-        compose:
-          - "{{ declare_vars }}"
-          - "{{ database }}"
-          - "{{ using_forbidden_database }}"
