--- examples/old/org_wide_policies/compliance/policybooks/check_pkg.yml
@@ -1,16 +1,0 @@
-# package-example.yml
----
-- name: Check for mysql package installation
-  hosts: localhost 
-  vars:
-    allowed_packages:
-      - "mysql-server"
-  policies:
-    - name: Check for package name
-      target: task
-      condition: input["ansible.builtin.package"].name not in allowed_packages
-      actions:
-        - deny:
-            msg: The package {{ input["ansible.builtin.package"].name }} is not allowed, allowed packages are one of {{ allowed_packages }}
-      tags:
-        - compliance
